Split wiznet driver into shared-bindings vs shared-module
